Refactored project layout into modular form
Refactored project layout into a more pythonic modular form.
