--- ccct/Makefile
@@ -1,5 +1,0 @@
-# SPDX-License-Identifier: MIT
-
-.PHONY: test
-test:
-	python3 -m unittest discover -v
