+++ pyproject.toml
@@ -1,0 +1,42 @@
+[project]
+name = "ccct"
+version = "0.0.1"
+description = "Credit Card Categorization Tool"
+readme = "README.md"
+license = "MIT"
+license-files = [ "LICENSE" ]
+requires-python = ">=3.11"
+authors = [{ name = "Chuck Wolber", email = "chuck@wolber.net" }]
+maintainers = [{ name = "Chuck Wolber", email = "chuck@wolber.net" }]
+keywords = ["QFX", "Financial", "CreditCard"]
+classifiers = [
+    "Intended Audience :: End Users/Desktop",
+    "Environment :: Console",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
+]
+dependencies = [
+    "ofxtools",
+    "google-api-python-client>=2.194.0",
+    "google-auth-httplib2>=0.3.1",
+    "google-auth-oauthlib>=1.3.1",
+    "jsonschema>=4.26.0",
+]
+
+[project.urls]
+Repository = "https://github.com/chuckwolber/cctools"
+Issues = "https://github.com/chuckwolber/cctools/issues"
+
+[project.scripts]
+ccct = "ccct.__main__:main"
+
+[build-system]
+requires = ["uv_build >= 0.11.7, <0.12"]
+build-backend = "uv_build"
+
+[dependency-groups]
+dev = ["pytest>=9.0.3"]
