+++ src/ccct/__main__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+#
+# SPDX-License-Identifier: MIT
+
+from ccct.ccct import console
+
+def main() -> None:
+    console()
+
+if __name__ == "__main__":
+    main()
