+++ src/ccct/ccct.py
@@ -8,6 +8,9 @@ import ofxtools
 import os.path
 import re
 
+from ccct.allocation import Allocation
+from ccct.transaction import Transaction
+
 from datetime import datetime
 from pathlib import Path
 from jsonschema import validate
@@ -29,218 +32,6 @@ ACCTTYPE = "CREDITLINE"
 SCOPES = ["https://www.googleapis.com/auth/spreadsheets"]
 TITLE = "CreditCardTransactions"
 
-class Allocation():
-    """Single-entry style expense allocation.
-
-    Categorize an expense single-entry style. No distinction is made between
-    the various double entry categories (asset, liability, capital, revenue,
-    expense) or types (CREDIT, DEBIT).
-
-    Amounts are presented to the user in terms they are used to seeing on a
-    traditional credit card statement. Negative values signify a reduction in
-    one's outstanding debt, such as a payment. Positive values signify an
-    increase in one's outstanding debt, usually stemming from a purchase or an
-    interest charge.
-
-    To avoid confusion, allocations can only be made using the same sign
-    presented to the user. For example, a transaction amount of $10 may not be
-    be allocated as $1000 to one category and -$990 to another category.
-    """
-    allocation: list
-    negative: bool
-    amount_orig: float
-    amount_curr: float
-
-    ALLOC_COLUMNS = None
-    ALLOC_COLUMNS_MAP = None
-
-    @classmethod
-    def init_cols(cls, alloc_columns: list):
-        """Initialize shared data."""
-        cls.ALLOC_COLUMNS = alloc_columns
-
-    @classmethod
-    def init_cols_map(cls, alloc_columns_map: list):
-        """Initialize shared data."""
-        cls.ALLOC_COLUMNS_MAP = alloc_columns_map
-
-    @classmethod
-    def __print_cols_map_for_key(cls, key: str):
-        print("\t" + key + ":\t" + cls.ALLOC_COLUMNS_MAP[key])
-
-    @classmethod
-    def print_cols_map(cls, key: str = None):
-        if cls.ALLOC_COLUMNS_MAP == None:
-            return
-        if key == None:
-            for key in cls.ALLOC_COLUMNS_MAP.keys():
-                cls.__print_cols_map_for_key(key)
-        else:
-            cls.__print_cols_map_for_key(key)
-
-    def __init__(self, amount: float):
-        if self.ALLOC_COLUMNS is None:
-            raise ValueError("Error: Allocation class not initialized. Call `initialize` first.")
-        if not isinstance(amount, float):
-            raise ValueError("Error: Amount is not a floating point value.")
-        self.allocation = [0.0] * len(self.ALLOC_COLUMNS)
-        self.amount_orig = float(-1*amount)
-        self.amount_curr = self.amount_orig
-        self.negative = False
-        if self.amount_orig < 0:
-            self.negative = True
-
-    def get_paired_allocations(self) -> str:
-        """Pair column categories.
-
-        Pair column categories with current amounts allocated to each category.
-        Omit the allocation amount if it is zero.
-        """
-        paired = [f"{a}({b})" if b != 0 else a for a, b in zip(self.ALLOC_COLUMNS, self.allocation)]
-        if self.ALLOC_COLUMNS_MAP != None:
-            paired.append('?')
-        return f"[{', '.join(paired)}]"
-
-    def to_list(self) -> list:
-        return ['' if x == 0 else x for x in self.allocation]
-
-    def get_allocations_sum(self) -> float:
-        return round(sum(self.allocation),2)
-
-    def __is_valid_category(self, category: str) -> bool:
-        if category in self.ALLOC_COLUMNS:
-            return True
-        return False
-
-    def allocate_amount(self, category: str, amount: str = None):
-        """Allocate an amount to a specific category.
-
-        Allocation amounts must be the same sign as the transaction amount.
-        The transaction amount can be applied to multiple categories until the
-        entire transaction amount is allocated. If the amount argument is
-        'None' the remaining unallocated amount is assumed. If a category has
-        an amount already allocated to it, the allocated value is added back to
-        the unallocated amount before a new allocation is applied.
-
-        Keyword arguments:
-        category -- the category to assign the amount
-        amount -- the currency value to assign to the category
-        """
-        if not self.__is_valid_category(category):
-            return
-        index = self.ALLOC_COLUMNS.index(category)
-
-        x = self.amount_curr
-        if amount != None:
-            try:
-                x = float(amount)
-            except ValueError:
-                return
-
-        # Allocation amount must have the same sign as the transaction amount.
-        if (self.negative and x > 0) or (not self.negative and x < 0):
-            return
-
-        # Reset over allocation to the current outstanding amount.
-        if self.negative and x < self.amount_curr:
-            x = self.amount_curr
-        elif not self.negative and x > self.amount_curr:
-            x = self.amount_curr
-
-        # Recover currently allocated amount before allocating a new amount.
-        self.amount_curr += self.allocation[index]
-
-        self.amount_curr = round(self.amount_curr - x, 2)
-        self.allocation[index] = x
-
-class Transaction():
-    """Transactions generally follow single-entry accounting rules.
-
-    A traditional human readable credit card statement tends to use a single
-    column to list charges as positive amounts and payments as negative amounts.
-    The total at the bottom is usually a positive value referring to what the
-    consumer owes the credit card company. This is an example of single-entry
-    accounting, where you are only concerned with cashflow and do not care to
-    make a distinction between various categories like revenue and expenses.
-
-    In contrast, OFX files (like QFX) are a bit closer to traditional double
-    entry accounting, except transaction amounts can be negative. True double
-    entry accounting tends to use positive transaction amounts and relies on the
-    meaning of CREDIT and DEBIT with respect to specific account type (e.g.
-    asset, liability, capital, revenue, expense) to track the relevant financial
-    picture.
-
-    Section 3.2.9.2 of the OFX v2.3 specification explains:
-
-        "An exception to the above rules are signage of the amount in statement
-         download transactions, wrapped within <STMTTRN></STMTTRN> tags. The
-         amounts in these transactions should be signed on the basis of how the
-         account is affected, e.g. a <TRNTYPE>DEBIT should have a negative
-         <TRNAMT> value."
-
-    Section 11.4.4.1 of the OFX v2.3 specifiction further clarifies:
-
-        "Transaction amounts are signed from the perspective of the customer.
-         For example, a credit card payment is positive while a credit card
-         purchase is negative."
-    """
-    TRANSACTION_COLUMNS = ["FITID", "DTPOSTED", "TRNTYPE", "TRNAMT", "NAME", "MEMO"]
-
-    fitid: str
-    dtposted: datetime
-    trntype: str
-    trnamt: float
-    name: str
-    memo: str
-    allocation: Allocation
-
-    def __init__(self, transaction):
-        self.fitid      = transaction.fitid
-        self.dtposted   = transaction.dtposted
-        self.trntype    = transaction.trntype
-        self.trnamt     = float(transaction.trnamt)
-        self.name       = transaction.name
-        self.memo       = transaction.memo
-        self.allocation = Allocation(self.trnamt)
-
-        if not isinstance(self.fitid, str):
-            raise ValueError("Error: FITID is not a string.")
-        if not isinstance(self.dtposted, datetime):
-            raise ValueError("Error: DTPOSTED is not a datetime object.")
-
-        # Currently only supporting CREDIT and DEBIT. Submit a bug if you find
-        # examples of others being used. OFX v2.3 section 11.4.4.3 lists the
-        # full set of valid values.
-        if self.trntype != "CREDIT" and self.trntype != "DEBIT":
-            raise ValueError("Error: Unknown transaction type.\n{}".format(str(transaction)))
-
-        # OFX v2.3 sections 3.2.9.2 and 11.4.4.1 specify sign value from the
-        # perspective of the customer.
-        if self.trntype == "CREDIT" and self.trnamt < 0:
-            raise ValueError("Error: Negative credit detected!\n{}".format(str(transaction)))
-        if self.trntype == "DEBIT"  and self.trnamt > 0:
-            raise ValueError("Error: Positive debit detected!\n{}".format(str(transaction)))
-
-        if not isinstance(self.name, str):
-            raise ValueError("Error: NAME is not a string.")
-
-    def to_list(self) -> list:
-        arr = [''] * len(self.TRANSACTION_COLUMNS)
-        arr[self.TRANSACTION_COLUMNS.index("FITID")]    = self.fitid
-        arr[self.TRANSACTION_COLUMNS.index("DTPOSTED")] = str(self.dtposted)
-        arr[self.TRANSACTION_COLUMNS.index("TRNTYPE")]  = self.trntype
-        arr[self.TRANSACTION_COLUMNS.index("TRNAMT")]   = self.trnamt
-        arr[self.TRANSACTION_COLUMNS.index("NAME")]     = self.name
-        arr[self.TRANSACTION_COLUMNS.index("MEMO")]     = self.memo
-        return arr
-
-    def print(self):
-        print("\tTID:\t{}".format(self.fitid))
-        print("\tDate:\t{}".format(self.dtposted.isoformat()))
-        print("\tType:\t{}".format(self.trntype))
-        print("\tAmount:\t{}".format(str(self.trnamt)))
-        print("\tName:\t{}".format(self.name))
-        print("\tMemo:\t{}".format(self.memo))
 
 #
 # ActionType functions called from _parse_args(). Neither the arguments nor the
@@ -761,7 +552,7 @@ def _write_ofx_transactions():
         print(f"An error occurred: {error}")
         return error
 
-def console_main():
+def console():
     _resolve_config()
     _parse_ofx_file()
     _get_google_creds()
