+++ src/ccct/config/ccct-example.config.json
