+++ test/assets/ccct.config.schema.json
