+++ test/assets/config.invalid-1.json
