+++ test/assets/config.invalid-2.json
