+++ test/assets/config.invalid-3.json
