+++ test/assets/config.schema.invalid-1.json
