+++ test/assets/config.valid.json
